Extract performance workspace context service
